prefer cookies.signed over cookies to prevent tampering
Automatically generates a signed representation of the cookie value and
verifies it when reading from the cookie again. This is useful for
creating cookies with values that the user is not supposed to change.
If a signed cookie was tampered with by the user (or a 3rd party), nil
will be returned.